fix
1.On first sign up users needs to be required to check a box that says
"I agree to the Terms and Condition and Privacy Policy" make sure both
are properly linked to the correct pages.
2.In the blue header area, to the right of the BuyAnyCoin Logo insert a
"Total Balance: $XX.XX" which will show the accumulated balance of all
the coins and USD in the users account. Additionally ensure this balance
is also shown in the "user" collection in the admin panel.
3./account - check boxes for news do not function. result does not save
after navigating away and then back to /accounts
4.password reset not working.